feat(feed): separate feed from profile
