+++ src/messages/ru.json
@@ -178,6 +178,7 @@
     },
     "menu": {
         "profile": "Профиль",
+        "feed": "Лента",
         "chat": "Чат",
         "community": "Сообщество"
     }
